refactor: split workers and restore graphs
