+++ client.py
@@ -11,10 +11,6 @@ try:
     import wmi
 except ImportError:
     wmi = None
-try:
-    import speedtest
-except ImportError:
-    speedtest = None
 import re
 import sys
 import time
@@ -23,9 +19,9 @@ from pathlib import Path
 from typing import List, Optional
 import shutil
 import subprocess
-import tempfile
-import locale
+
 import psutil
+from client.worker import run_speedtest, run_diagnostics, submit
 # Кэш для вычисления CPU без задержки
 PROC_CACHE: dict[int, tuple[float, float]] = {}
 GPU_VENDOR: str | None = None
@@ -596,61 +592,15 @@ def gather_metrics(full: bool = False) -> dict:
     return data
 
 
-def _subprocess_flags() -> int:
-    """Return flags for subprocess to avoid extra windows on Windows."""
-    if platform.system() == "Windows":
-        return subprocess.CREATE_NO_WINDOW
-    return 0
-
-
-def run_speedtest() -> tuple[float | None, float | None, float | None]:
-    """Run a network speed test using any available backend."""
-    try:
-        if speedtest:
-            try:
-                # PyInstaller --noconsole workaround: disable library prints
-                speedtest.printer = lambda *a, **k: None
-            except Exception:
-                pass
-            st = speedtest.Speedtest(secure=True)
-            st.get_best_server()
-            dl = st.download() / 1e6
-            ul = st.upload() / 1e6
-            return dl, ul, st.results.ping
-
-        import shutil, subprocess, json
-
-        for prog in (
-            "speedtest",
-            "speedtest-cli",
-            "speedtest.exe",
-            "speedtest-cli.exe",
-        ):
-            path = shutil.which(prog)
-            if not path:
-                continue
-            out = subprocess.check_output(
-                [path, "--format=json"],
-                text=True,
-                timeout=120,
-                creationflags=_subprocess_flags(),
-            )
-            data = json.loads(out)
-            dl = data["download"]["bandwidth"] * 8 / 1e6
-            ul = data["upload"]["bandwidth"] * 8 / 1e6
-            ping = data["ping"]["latency"]
-            return dl, ul, ping
-    except Exception as exc:
-        log.error("speedtest failed: %s", exc)
-    return None, None, None
 # ---------- async speedtest helper ----------
 speedtest_running = False      # флаг «тест уже идёт»
+diag_running = False
 
 def _speedtest_job():
     global speedtest_running
     try:
         push_text("⏳ Тестируем скорость…")
-        dl, ul, ping = run_speedtest()
+        dl, ul, ping = submit(run_speedtest).result()
         if dl is not None:
             push_text(
                 f"💨 Speedtest:\n"
@@ -663,62 +613,6 @@ def _speedtest_job():
     finally:
         speedtest_running = False
 
-# ---------- diagnostics helper ----------
-diag_running = False
-
-def run_diagnostics() -> str | None:
-    """Collect diagnostics data using available system tools."""
-    try:
-        if platform.system() == "Windows":
-            dxdiag = shutil.which("dxdiag") or shutil.which("dxdiag.exe")
-            if dxdiag:
-                tmp = Path(tempfile.gettempdir()) / "dxdiag.txt"
-                cmd = [dxdiag, "/dontskip", "/whql:off", "/t", str(tmp)]
-                subprocess.run(
-                    cmd,
-                    check=True,
-                    timeout=120,
-                    creationflags=_subprocess_flags(),
-                )
-                try:
-                    return tmp.read_text(encoding="utf-16")
-                except UnicodeError as exc:
-                    log.warning("dxdiag UTF-16 parse failed: %s", exc)
-                    enc = locale.getpreferredencoding(False)
-                    try:
-                        return tmp.read_text(encoding=enc, errors="ignore")
-                    except UnicodeError:
-                        return tmp.read_text(encoding="utf-8", errors="ignore")
-            sysinfo = shutil.which("systeminfo") or shutil.which("systeminfo.exe")
-            if sysinfo:
-                out = subprocess.check_output(
-                    [sysinfo],
-                    text=True,
-                    timeout=120,
-                    errors="ignore",
-                    creationflags=_subprocess_flags(),
-                )
-                return out
-
-        if shutil.which("inxi"):
-            out = subprocess.check_output(
-                ["inxi", "-F"],
-                text=True,
-                timeout=120,
-                creationflags=_subprocess_flags(),
-            )
-            return out
-        if shutil.which("lshw"):
-            out = subprocess.check_output(
-                ["lshw", "-short"],
-                text=True,
-                timeout=120,
-                creationflags=_subprocess_flags(),
-            )
-            return out
-    except Exception as exc:
-        log.error("diagnostics failed: %s", exc)
-    return None
 
 def push_diag(txt: str, ok: bool = True):
     """Send diagnostics result to the server."""
@@ -736,7 +630,7 @@ def _diag_job():
     global diag_running
     try:
         push_text("⏳ Собираем диагностику…")
-        out = run_diagnostics()
+        out = submit(run_diagnostics).result()
         if out:
             push_diag(out, ok=True)
         else:
